[#113313657] Refactor Dashboard Components
Merge the Dashboard index component and the DocList component into one
